fix: types for canvasdrawlines and lines
